Move some 'models' in a new 'ViewModels.swift' file
